Begin to add UI system
Revamped the font loading engine to always just use the highest
resolution available, rather than offering those of different sizes. We
can just scale it down, if needed (SDL_TTF will handle it for us,
anyway).

Revamped the UI controls system to do away with the functions that
handle all control types and just have a collection of functions for
each type. This makes the code cleaner and is more fitted to my final
vision of what the menu system will work like.

Added the UI code for labels.