Add UI button control
The button will automatically detect focus on a per-frame basis, and
toggle to another caller-defined theme. This theme will always have a
white background, as defined by the texture file "fbutton.png".

Additionally, the button will call hooks (with automatic debouncing)
when focus is gained or when the button is clicked. This will happen
through function pointers. The event will be ignored if the function
pointer is not set.

This commit adds two new resource dependencies in the res/ui directory.