Add main and testing menus to menu system
Testing menu just emulates the behaviour of the existing testing system
that was just dumped in the main function. The init routine will just
put the testing menu in the currentMenu pointer and set the state to
fsMenu.

So, essentially, we have the same program but a bit cleaner and ready to
be made into a full game.